Don't rely on a custom root tag for the layout xml.
This commit integrates a suggestion by Jerome P to use <merge/> as the root
element for my view group layout. It means that the ContactView class itself
is responsible for inflating the layout and that the ListAdapter needs only
to new ContactView(). This creates a better kind of coupling between code
and XML.